+++ src/main/java/bc1/gream/domain/order/mapper/BuyMapper.java
@@ -1,7 +1,7 @@
-package bc1.gream.domain.buy.mapper;
+package bc1.gream.domain.order.mapper;
 
-import bc1.gream.domain.buy.dto.response.BuyBidResponseDto;
-import bc1.gream.domain.buy.entity.Buy;
+import bc1.gream.domain.order.dto.response.BuyBidResponseDto;
+import bc1.gream.domain.order.entity.Buy;
 import bc1.gream.domain.product.dto.TradeResponseDto;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
