+++ src/main/java/bc1/gream/domain/order/repository/SellRepositoryCustom.java
@@ -1,7 +1,7 @@
-package bc1.gream.domain.sell.repository;
+package bc1.gream.domain.order.repository;
 
+import bc1.gream.domain.order.entity.Sell;
 import bc1.gream.domain.product.entity.Product;
-import bc1.gream.domain.sell.entity.Sell;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
