+++ src/main/java/bc1/gream/domain/order/service/BuyService.java
@@ -1,12 +1,15 @@
-package bc1.gream.domain.buy.service;
+package bc1.gream.domain.order.service;
 
+import static bc1.gream.global.common.ResultCase.BUY_BID_PRODUCT_NOT_FOUND;
+import static bc1.gream.global.common.ResultCase.NOT_AUTHORIZED;
 import static bc1.gream.global.common.ResultCase.PRODUCT_NOT_FOUND;
 
-import bc1.gream.domain.buy.dto.request.BuyBidRequestDto;
-import bc1.gream.domain.buy.dto.response.BuyBidResponseDto;
-import bc1.gream.domain.buy.entity.Buy;
-import bc1.gream.domain.buy.mapper.BuyMapper;
-import bc1.gream.domain.buy.repository.BuyRepository;
+import bc1.gream.domain.order.dto.request.BuyBidRequestDto;
+import bc1.gream.domain.order.dto.response.BuyBidResponseDto;
+import bc1.gream.domain.order.dto.response.BuyCancelBidResponseDto;
+import bc1.gream.domain.order.entity.Buy;
+import bc1.gream.domain.order.mapper.BuyMapper;
+import bc1.gream.domain.order.repository.BuyRepository;
 import bc1.gream.domain.product.entity.Product;
 import bc1.gream.domain.product.repository.ProductRepository;
 import bc1.gream.domain.user.entity.User;
@@ -47,6 +50,19 @@ public class BuyService {
         return BuyMapper.INSTANCE.toBuyBidResponseDto(savedBuy);
     }
 
+
+    public BuyCancelBidResponseDto buyCancelBid(User user, Long buyId) {
+        Buy buyBid = findBuyById(buyId);
+
+        if (isNotBuyerLoggedInUser(buyBid, user)) {
+            throw new GlobalException(NOT_AUTHORIZED);
+        }
+
+        buyRepository.delete(buyBid);
+
+        return new BuyCancelBidResponseDto();
+    }
+
     private Integer getPeriod(Integer period) {
         return Objects.requireNonNullElse(period, 7);
     }
@@ -56,6 +72,17 @@ public class BuyService {
             () -> new GlobalException(PRODUCT_NOT_FOUND)
         );
     }
+
+    protected Buy findBuyById(Long buyId) {
+        return buyRepository.findById(buyId).orElseThrow(
+            () -> new GlobalException(BUY_BID_PRODUCT_NOT_FOUND)
+        );
+    }
+
+    private boolean isNotBuyerLoggedInUser(Buy buy, User user) {
+        return !buy.getUser().getLoginId().equals(user.getLoginId());
+    }
+
 
     /**
      * Product에 대한 구매입찰가 내역 페이징 조회
