+++ src/main/java/bc1/gream/domain/order/service/SellService.java
@@ -8,6 +8,7 @@ import bc1.gream.domain.order.dto.request.SellBidRequestDto;
 import bc1.gream.domain.order.dto.response.SellBidResponseDto;
 import bc1.gream.domain.order.dto.response.SellCancelBidResponseDto;
 import bc1.gream.domain.order.entity.Sell;
+import bc1.gream.domain.order.mapper.SellMapper;
 import bc1.gream.domain.order.repository.SellRepository;
 import bc1.gream.domain.product.entity.Product;
 import bc1.gream.domain.product.repository.ProductRepository;
@@ -18,7 +19,10 @@ import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.util.Objects;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -46,7 +50,7 @@ public class SellService {
 
         Sell savedSell = sellRepository.save(sell);
 
-        return SellServiceMapper.INSTANCE.toSellBidResponseDto(savedSell);
+        return SellMapper.INSTANCE.toSellBidResponseDto(savedSell);
     }
 
 
@@ -80,5 +84,18 @@ public class SellService {
 
     private boolean isNotSellerLoggedInUser(Sell sell, User user) {
         return !sell.getUser().getLoginId().equals(user.getLoginId());
+    }
+
+
+    /**
+     * Product에 대한 판매입찰가 내역 페이징 조회
+     *
+     * @param product  이모티콘 상품
+     * @param pageable 페이징 요청 데이터
+     * @return 판매입찰가 내역 페이징 데이터
+     */
+    @Transactional(readOnly = true)
+    public Page<Sell> findAllSellBidsOf(Product product, Pageable pageable) {
+        return sellRepository.findAllPricesOf(product, pageable);
     }
 }
