+++ src/main/java/bc1/gream/domain/product/service/SellOrderQueryService.java
@@ -1,9 +1,9 @@
 package bc1.gream.domain.product.service;
 
+import bc1.gream.domain.order.entity.Sell;
+import bc1.gream.domain.order.service.SellService;
 import bc1.gream.domain.product.entity.Product;
 import bc1.gream.domain.product.service.query.ProductQueryService;
-import bc1.gream.domain.sell.entity.Sell;
-import bc1.gream.domain.sell.service.SellService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
